Hidde delta when there is no previous or current

+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/RevenueStat.kt
@@ -2,8 +2,8 @@ package com.woocommerce.android.model
 
 data class RevenueStat(
     val totalValue: Double,
-    val totalDelta: Int,
+    val totalDelta: Double,
     val netValue: Double,
-    val netDelta: Int,
+    val netDelta: Double,
     val currencyCode: String?
 )
